Minor cleanups relating to size_t being unsigned
In order to use -1 as a sentinel value, we should explicitly cast
to make it clear what we are doing.  It might be better to use
a less convoluted sentinel value such as SIZE_T_MAX, though.

Additionally, since size_t is unsigned and at least as wide as
an int, a loop with int index variable that compares against a
size_t for its termination check could become an infinite loop.
Make the loop index size_t for consistency.

+++ src/plugins/kdb/db2/libdb2/hash/hash_bigkey.c
@@ -277,7 +277,7 @@ __big_keydata(hashp, pagep, key, val, ndx)
 	key->data = hashp->bigkey_buf;
 	__put_page(hashp, key_pagep, A_RAW, 0);
 
-	if (key->size == -1)
+	if (key->size == (size_t)-1)
 		return (-1);
 
 	/* Create an item_info to direct __big_return to the beginning pgno. */
